Major Streaming Branch Upgrades: GPT4, No more rumps, no more pygame, optional APIs, dynamic icon
GPT 4 was added as a new option for the chatbot. The app now uses PyQT6 instead of rumps because it will
give us more options in the future, possibly will allow for cross platform compatibility..The app now uses
a new module based soley on pyaudio removing the need for pygame. The app no longer requires all the APIs
I recommended be used (excluding openai of course) and that means each now has a free alternative laid out
in the readme. The app now has a dynamic icon that changes based on the state of the app, which is critical
for understanding what is going on in the app. Finally, the app is not constantly leaking semaphores.
For the future, this branch still hasn't been package tested and the app still needs to be tested in practice
but after a few good days of use, with minor tweaks I think this branch is ready to be merged into the main branch.